added mltiple algorithm and structured the code

+++ src/main/java/com/natwest/primeNumberService/PrimeNumber.java
@@ -3,17 +3,16 @@ package com.natwest.primeNumberService;
 import java.util.List;
 
 
-
 public class PrimeNumber {
 
-	private final Long initial;
+	private final Integer initial;
 	private final List primes;
 	
-	public PrimeNumber(Long initial, List primes) {
+	public PrimeNumber(Integer initial, List primes) {
 		this.initial = initial;
 		this.primes = primes;
 	}
-	public Long getInitial() {
+	public Integer getInitial() {
 		return initial;
 	}
 	public List getPrimes() {
